week 2

+++ ContosoUniversity/Controllers/StudentsController.cs
@@ -34,7 +34,11 @@ namespace ContosoUniversity.Controllers
             }
 
             var student = await _context.Students
-                .SingleOrDefaultAsync(m => m.ID == id);
+               // .SingleOrDefaultAsync(m => m.ID == id);
+               .Include(S => S.Enrollments)
+               .ThenInclude(e => e.Course)
+               .AsNoTracking()
+               .FirstOrDefaultAsync(m => m.ID == id);
             if (student == null)
             {
                 return NotFound();
@@ -54,18 +58,32 @@ namespace ContosoUniversity.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ID,LastName,FirstMidName,EnrollmentDate")] Student student)
+        public async Task<IActionResult> Create([Bind("EnrollmentDate,FirstMidName,LastName")] Student student)
         {
-            if (ModelState.IsValid)
+            try
             {
-                _context.Add(student);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(student);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
             }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error(uncomment ex variable name and write log.
+                ModelState.AddModelError("", "Unable to save changes." + "Try again, and if the problem persists" + "see your administrator.");
+            }
+            //    if (ModelState.IsValid)
+            //    {
+            //        _context.Add(student);
+            //        await _context.SaveChangesAsync();
+            //        return RedirectToAction(nameof(Index));
+            //    }
             return View(student);
         }
 
-        // GET: Students/Edit/5
+        //// GET: Students/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -84,39 +102,96 @@ namespace ContosoUniversity.Controllers
         // POST: Students/Edit/5
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ID,LastName,FirstMidName,EnrollmentDate")] Student student)
+        public async Task<IActionResult> EditPost(int? id)
         {
-            if (id != student.ID)
+            if (id == null)
             {
                 return NotFound();
             }
-
-            if (ModelState.IsValid)
+            var studentToUpdate = await _context.Students.FirstOrDefaultAsync(s => s.ID == id);
+            if (await TryUpdateModelAsync<Student>(studentToUpdate, "",
+                s => s.FirstMidName, s => s.LastName, StudentToUpdate => studentToUpdate.EnrollmentDate))
             {
                 try
                 {
-                    _context.Update(student);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
+                catch (DbUpdateException /*ex*/)
                 {
-                    if (!StudentExists(student.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    //Log the error(Uncomment ex variable name and write a log.
+                    ModelState.AddModelError("", "Unable to save changes" + "Try again,If the problem persists" + "see your administrator.");
                 }
-                return RedirectToAction(nameof(Index));
             }
-            return View(student);
+            return View(studentToUpdate);
+
+
         }
 
+
+
+        /*   
+       [HttpPost]
+       [ValidateAntiForgeryToken]
+       public async Task<IActionResult> Edit(int id, [Bind("ID,LastName,FirstMidName,EnrollmentDate")] Student student)
+       {
+           if (id != student.ID)
+           {
+               return NotFound();
+           }
+
+           if (ModelState.IsValid)
+           {
+               try
+               {
+                   _context.Update(student);
+                   await _context.SaveChangesAsync();
+               }
+               catch (DbUpdateConcurrencyException)
+               {
+                   if (!StudentExists(student.ID))
+                   {
+                       return NotFound();
+                   }
+                   else
+                   {
+                       throw;
+                   }
+               }
+               return RedirectToAction(nameof(Index));
+           }
+           return View(student);
+       }*/
+
         // GET: Students/Delete/5
+        public async Task<IActionResult> Delete(int? id, bool? saveChangesError = false)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+            var student = await _context.Students
+                .AsNoTracking()
+                .FirstOrDefaultAsync(m => m.ID == id);
+            if (student == null)
+            {
+                return NotFound();
+            }
+            if (saveChangesError.GetValueOrDefault())
+            {
+                ViewData["Error Message"] =
+                    "Delete Failed. Try again, and if the problem persists" + "See your system administrator";
+            }
+            return View(student);
+        }      
+    
+
+
+
+
+        /*
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -133,21 +208,47 @@ namespace ContosoUniversity.Controllers
 
             return View(student);
         }
-
+        */
         // POST: Students/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
-            _context.Students.Remove(student);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var student = await _context.Students.FindAsync(id);
+            if(student==null)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+            try
+            {
+                _context.Students.Remove(student);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch  (DbUpdateException /*eg */)
+            {
+                //Log the error(Uncomment ex variable and write a log
+                return RedirectToAction(nameof(Delete), new { id = id, saveChangesError = true });
+            }
         }
 
-        private bool StudentExists(int id)
-        {
-            return _context.Students.Any(e => e.ID == id);
-        }
+
+
+
+
+        //[HttpPost, ActionName("Delete")]
+        //[ValidateAntiForgeryToken]
+        //public async Task<IActionResult> DeleteConfirmed(int id)
+        //{
+        //    var student = await _context.Students.SingleOrDefaultAsync(m => m.ID == id);
+        //    _context.Students.Remove(student);
+        //    await _context.SaveChangesAsync();
+        //    return RedirectToAction(nameof(Index));
+        //}
+
+        //private bool StudentExists(int id)
+        //{
+        //    return _context.Students.Any(e => e.ID == id);
+        //}
     }
 }
